box: move port and region ffi cdef from schema to utils
We use them not only in box/schema.lua but also in box/read_view.lua.
The problem is that box/schema.lua isn't loaded in application threads
while box/read_view.lua is going to be. So let's move the port and
region FFI C API from box/schema.lua to box/utils.lua, which is loaded
in all threads.

Needed for #12951

NO_DOC=refactoring
NO_TEST=refactoring
NO_CHANGELOG=refactoring
